[iOS] Fix setting initial CurrentItem
