+++ Xamarin.Forms.Platform.iOS/Xamarin.Forms.Platform.iOS.csproj
@@ -257,6 +257,7 @@
     <Compile Include="Renderers\IndicatorViewRenderer.cs" />
     <Compile Include="Renderers\SwipeViewRenderer.cs" />
     <Compile Include="Extensions\UIApplicationExtensions.cs" />
+    <Compile Include="Extensions\CollectionViewExtensions.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Resources\StringResources.ar.resx" />
